refactor(server): extract updateSingleBusiness helper
Move the per-business update body of the updateBusiness mutation (core
financial-entity fields, business fields, suggestions, tax-category
match, green-invoice sync) into a reusable updateSingleBusiness helper
that returns the refreshed business and throws on failure. The
updateBusiness resolver now wraps the helper and maps thrown errors to
CommonError, preserving its observable behavior. No schema change; this
sets up batchUpdateBusinesses to reuse the same logic.
